+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- AuthN user password expiration support.
+
 ## [3.10.1] - 2024-06-24
 
 ### Fixed